feat: updated Readme & fixed first time setup steps.
