+++ .env.example
@@ -1,0 +1,1 @@
+ALCHEMY_API_KEY="YOUR_API_KEY"
