+++ packages/backend/src/snapshot.ts
@@ -8,22 +8,8 @@ import { getAllLogs } from "./events";
 import { fetchAuctions, fetchAuctionsResponse } from "./propHouse";
 import { z } from "zod";
 import { nounsDao, nounsToken } from "./contracts";
+import { TypedInterface, ContractInstance} from "./utils/contract";
 
-export interface TypedInterface extends ethers.utils.Interface {
-  events: Record<string, ethers.utils.EventFragment<Record<string, any>>>;
-}
-
-export function makeContractInstance<InterfaceType extends TypedInterface>(
-  t: ContractInstance<InterfaceType>
-): ContractInstance<InterfaceType> {
-  return t;
-}
-
-type ContractInstance<InterfaceType extends TypedInterface> = {
-  iface: InterfaceType;
-  address: string;
-  startingBlock: number;
-};
 type ReducerDefinition<
   InterfaceType extends TypedInterface,
   Accumulator,
