+++ packages/backend/src/utils/contract.ts
@@ -1,0 +1,17 @@
+import { ethers } from "ethers";
+
+export interface TypedInterface extends ethers.utils.Interface {
+  events: Record<string, ethers.utils.EventFragment<Record<string, any>>>;
+}
+
+export function makeContractInstance<InterfaceType extends TypedInterface>(
+  t: ContractInstance<InterfaceType>
+): ContractInstance<InterfaceType> {
+  return t;
+}
+
+export type ContractInstance<InterfaceType extends TypedInterface> = {
+  iface: InterfaceType;
+  address: string;
+  startingBlock: number;
+};
